add separate server/client helper types
